refactor: trying out a different organization
